+++ bracket-bevy/src/consoles/sparse_console/back_end/mod.rs
@@ -19,5 +19,5 @@ pub(crate) trait SparseConsoleBackend: Sync + Send {
     ) -> Handle<Mesh>;
     fn spawn(&self, commands: &mut Commands, material: Handle<ColorMaterial>, idx: usize);
     fn get_pixel_size(&self) -> (f32, f32);
-    fn resize(&mut self, available_size: &(f32, f32)) -> (usize, usize);
+    fn resize(&mut self, available_size: &(f32, f32)) -> (i32, i32);
 }
